requirements.txt
cupy-cuda100==7.1.1
numpy==1.18.1
scipy==1.3.2
scikit-learn==0.22
pandas==0.25.3